requirements.txt
amqp==5.2.0
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
astroid==3.0.1
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.5
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.7
Django==4.2.7
django-allauth==0.58.2
django-appconf==1.0.6
django-auditlog==2.3.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-compressor==4.4
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.4
django-import-export==3.3.3
django-model-utils==4.3.1
django-redis==5.4.0
django-tenants==3.5.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
flower==2.0.1
h11==0.14.0
httptools==0.6.1
humanize==4.9.0
idna==3.5
inflection==0.5.1
isort==5.12.0
kombu==5.3.4
MarkupPy==1.14
mccabe==0.7.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
platformdirs==4.0.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
pylint==3.0.2
pylint-celery==0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
rcssmin==1.1.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tablib==3.5.0
tenant-schemas-celery==2.2.0
text-unidecode==1.3
tomlkit==0.12.3
tornado==6.3.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0